时间改成Integer

+++ src/main/java/com/newframe/dto/account/response/AccountFundingDTO.java
@@ -45,9 +45,9 @@ public class AccountFundingDTO {
     /**
      * ctime
      */
-    private Long ctime;
+    private Integer ctime;
     /**
      * utime
      */
-    private Long utime;
+    private Integer utime;
 }
